laacsite/requirements.txt
asgiref==3.8.1
dataclasses==0.6
dj-database-url==2.3.0
Django==5.1.5
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.7.0
psycopg2-binary==2.9.10
pymongo==4.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
